cs: only try cleaning old entity locks
EntityLockCleanerService will only query entity locks
which are older than the given duration.
Caller this service can set this duration at creation time.

+++ java/cloud-store/src/main/java/com/vmware/photon/controller/cloudstore/dcp/task/EntityLockCleanerService.java
@@ -51,6 +51,7 @@ import java.util.Map;
 public class EntityLockCleanerService extends StatefulService {
 
   public static final Integer ENTITY_LOCK_DEFAULT_PAGE_LIMIT = 1000;
+  private static final String DOCUMENT_UPDATE_TIME_MICROS = "documentUpdateTimeMicros";
 
   public EntityLockCleanerService() {
     super(State.class);
@@ -156,7 +157,7 @@ public class EntityLockCleanerService extends StatefulService {
       switch (current.taskState.stage) {
         case STARTED:
           final State finishPatch = new State();
-          processUnreleasedEntityLocks(finishPatch);
+          processUnreleasedEntityLocks(finishPatch, current);
           break;
 
         case FAILED:
@@ -181,10 +182,10 @@ public class EntityLockCleanerService extends StatefulService {
    *
    * @param finishPatch
    */
-  private void processUnreleasedEntityLocks(final State finishPatch) {
+  private void processUnreleasedEntityLocks(final State finishPatch, final State current) {
     Operation queryEntityLocksPagination = Operation
         .createPost(UriUtils.buildUri(getHost(), ServiceUriPaths.CORE_LOCAL_QUERY_TASKS))
-        .setBody(buildEntityLockQuery());
+        .setBody(buildEntityLockQuery(current));
 
     Operation getFirstPageOfEntityLocks = Operation.createGet(null);
 
@@ -211,28 +212,43 @@ public class EntityLockCleanerService extends StatefulService {
             failTask(throwable.values().iterator().next());
             return;
           }
-          Operation op = ops.get(getFirstPageOfEntityLocks.getId());
-          List<EntityLockService.State> entityLockList =
-              parseEntityLockQueryResults(op.getBody(QueryTask.class));
+          if (getFirstPageOfEntityLocks.getUri() != null) {
+            Operation op = ops.get(getFirstPageOfEntityLocks.getId());
 
-          if (entityLockList.size() == 0) {
-            ServiceUtils.logInfo(EntityLockCleanerService.this, "No entityLocks found.");
-            finishTask(finishPatch);
-            return;
+            List<EntityLockService.State> entityLockList =
+                parseEntityLockQueryResults(op.getBody(QueryTask.class));
+
+            if (entityLockList.size() == 0) {
+              ServiceUtils.logInfo(EntityLockCleanerService.this, "No entityLocks found.");
+              finishTask(finishPatch);
+              return;
+            }
+
+            deleteUnreleasedEntityLocks(finishPatch, entityLockList);
           }
-          deleteUnreleasedEntityLocks(finishPatch, entityLockList);
         })
         .sendWith(this);
   }
 
-  private QueryTask buildEntityLockQuery() {
+  private QueryTask buildEntityLockQuery(final State current) {
+    Long duration = Utils.getNowMicrosUtc() - current.entityLockDeleteWatermarkTime;
+
     QueryTask.Query kindClause = new QueryTask.Query()
         .setTermPropertyName(ServiceDocument.FIELD_NAME_KIND)
         .setTermMatchValue(Utils.buildKind(EntityLockService.State.class));
 
     QueryTask.QuerySpecification querySpec = new QueryTask.QuerySpecification();
 
-    querySpec.query = kindClause;
+    QueryTask.NumericRange range = QueryTask.NumericRange.createLessThanRange(duration);
+    range.precisionStep = Integer.MAX_VALUE;
+    QueryTask.Query timeClause = new QueryTask.Query()
+        .setTermPropertyName(DOCUMENT_UPDATE_TIME_MICROS)
+        .setNumericRange(range);
+
+    querySpec.query
+        .addBooleanClause(kindClause)
+        .addBooleanClause(timeClause);
+
     querySpec.options = EnumSet.of(QueryTask.QuerySpecification.QueryOption.EXPAND_CONTENT);
     querySpec.resultLimit = ENTITY_LOCK_DEFAULT_PAGE_LIMIT;
     return QueryTask.create(querySpec).setDirect(true);
@@ -320,9 +336,11 @@ public class EntityLockCleanerService extends StatefulService {
   private List<EntityLockService.State> parseEntityLockQueryResults(QueryTask result) {
     List<EntityLockService.State> entityLockList = new LinkedList<>();
 
-    for (Map.Entry<String, Object> doc : result.results.documents.entrySet()) {
-      entityLockList.add(
-          Utils.fromJson(doc.getValue(), EntityLockService.State.class));
+    if (result.results.documentCount > 0) {
+      for (Map.Entry<String, Object> doc : result.results.documents.entrySet()) {
+        entityLockList.add(
+            Utils.fromJson(doc.getValue(), EntityLockService.State.class));
+      }
     }
 
     return entityLockList;
@@ -434,5 +452,12 @@ public class EntityLockCleanerService extends StatefulService {
      */
     @DefaultBoolean(value = false)
     public Boolean isSelfProgressionDisabled;
+
+    /**
+     * Duration that controls how old the entity locks should be for cleaning.
+     */
+    @DefaultInteger(value = 0)
+    public Integer entityLockDeleteWatermarkTime;
+
   }
 }
